+++ changelog.md
@@ -15,6 +15,8 @@ Codion Change Log
 - AbstractComponentBuilder now supports multiple linked values.
 ### is.codion.swing.framework.ui
 - EntityDialogs.DefaultEditAttributeDialogBuilder bug fixed, didn't take null into account when counting distinct values.
+- EntityTablePanel, TableControl.EDIT_SELECTED renamed EDIT_VALUE and now associated with an action for editing the selected items.
+- EntityTablePanel.Config.editAttributeSelection() and includeEditValueControl() added.
 
 ## 0.17.41
 ### is.codion.common.core